split up widgets
